Vary automatic references to related ents
When adding multiple ents of the same type, don't have them all point at the
same related ents, but instead cycle through the ents that are already in the
DB.

Note that for this to work the referenced ents need to be in the DB by the time
the referencing ent is added, which can be ensured through progressive
construction (i.e. calling `add-ents` multiple times).